extract Pixel World visual fixture data
